chore: move @sidebar slot _components to root dir _components folder
