+++ src/app/@sidebar/ask-anything/page.tsx
@@ -1,4 +1,4 @@
-import GeneralSideBar from "../_components/GeneralSidebar";
+import GeneralSideBar from "../../_components/GeneralSidebar"
 
 export default function Page() {
   return <GeneralSideBar />
