--- src/app/@sidebar/default.tsx
@@ -1,5 +1,0 @@
-export default function Default() {
-  return (
-    <></>
-  );
-}
